LPD-37555 - extract clickActionInRow into a util to be imported
